requirements.txt
Babel==2.6.0
Django==2.1.4
django-js-asset==1.1.0
django-mptt==0.9.1
django-phonenumber-field==2.1.0
django-river==0.10.0
factory-boy==2.11.1
Faker==1.0.0
mock==2.0.0
pbr==5.1.1
phonenumbers==8.10.2
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
